Update to About.jsx
Edit component was extracted from the About file into a new file an dAbout has been successfully imported into channel details and settings
